+++ Compose-Flight-Search-App/core/testing/src/main/kotlin/com/example/flightsearchapp/core/testing/fake/di/TestAirportsFtsRepositoryModule.kt
@@ -1,8 +1,8 @@
-package com.example.flightsearchapp.testing.di
+package com.example.flightsearchapp.core.testing.fake.di
 
-import com.example.flightsearchapp.data.AirportsFtsRepository
-import com.example.flightsearchapp.di.AirportsFtsRepositoryModule
-import com.example.flightsearchapp.testing.repository.AirportsFtsFakeRepository
+import com.example.flightsearchapp.core.data.di.AirportsFtsRepositoryModule
+import com.example.flightsearchapp.core.data.repository.AirportsFtsRepository
+import com.example.flightsearchapp.core.testing.fake.repository.AirportsFtsFakeRepository
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.components.SingletonComponent
