+++ Compose-Flight-Search-App/core/testing/src/main/kotlin/com/example/flightsearchapp/core/testing/fake/di/TestSearchTextDataStoreModule.kt
@@ -1,25 +1,24 @@
-package com.example.flightsearchapp.testing.di
+package com.example.flightsearchapp.core.testing.fake.di
 
 import android.content.Context
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.PreferenceDataStoreFactory
 import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.preferencesDataStoreFile
-import com.example.flightsearchapp.di.DataStoreModule
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
 import dagger.hilt.testing.TestInstallIn
-import javax.inject.Singleton
 import kotlinx.coroutines.test.TestScope
+import javax.inject.Singleton
 
 private const val TEST_SEARCH_TEXT = "test_search_text"
 
 @Module
 @TestInstallIn(
     components = [SingletonComponent::class],
-    replaces = [DataStoreModule::class],
+    replaces = [com.example.flightsearchapp.core.data.di.DataStoreModule::class],
 )
 object TestSearchTextDataStoreModule {
     @Provides
